:hammer: refactor: 통계에서 총 수익금을 구할 때 for 대신 stream 사용 & 당첨금 prize 를 long 으로 변경

+++ src/main/java/lotto/domain/WinningCriteria.java
@@ -13,7 +13,7 @@ public enum WinningCriteria {
 
     private final int matchCount;
     private final Set<Boolean> bonusResults;
-    private final int prize;
+    private final long prize;
 
     WinningCriteria(final int matchCount, final int prize) {
         this(matchCount, Set.of(false, true), prize);
@@ -39,7 +39,7 @@ public enum WinningCriteria {
         return this.matchCount;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return this.prize;
     }
 }
